Add MQTT message
