+++ CMakeLists.txt
@@ -28,6 +28,7 @@ target_sources(warp
   PRIVATE
     src/warp/warp.cpp
     src/warp/mqtt/codec.cpp
+    src/warp/mqtt/message.cpp
     src/warp/mqtt/server.cpp
 )
