+++ src/warp/mqtt/message.cpp
@@ -1,0 +1,5 @@
+#include "warp/mqtt/message.h"
+
+namespace warp::mqtt {
+// empty
+}  // namespace warp::mqtt
